refactor: move run cli command function to default doc tool class
